// src/openapi3/aether/2.1.0/models/site.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertyEnterpriseId } from './additional-property-enterprise-id';
import { SiteConnectivityService } from './site-connectivity-service';
import { SiteDeviceGroupList } from './site-device-group-list';
import { SiteDeviceList } from './site-device-list';
import { SiteImsiDefinition } from './site-imsi-definition';
import { SiteIpDomainList } from './site-ip-domain-list';
import { SiteMonitoring } from './site-monitoring';
import { SiteSimCardList } from './site-sim-card-list';
import { SiteSliceList } from './site-slice-list';
import { SiteSmallCellList } from './site-small-cell-list';
import { SiteUpfList } from './site-upf-list';

/**
 * List of site (single)
 */
export interface Site {
  'connectivity-service'?: SiteConnectivityService;

  /**
   * long description field
   */
  description?: string;
  device?: SiteDeviceList;
  'device-group'?: SiteDeviceGroupList;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;
  'imsi-definition'?: SiteImsiDefinition;
  'ip-domain'?: SiteIpDomainList;
  monitoring?: SiteMonitoring;
  'sim-card'?: SiteSimCardList;

  /**
   * ID for this site.
   */
  'site-id': string;
  slice?: SiteSliceList;
  'small-cell'?: SiteSmallCellList;
  upf?: SiteUpfList;

  [key: string]: AdditionalPropertyEnterpriseId | SiteConnectivityService | SiteDeviceGroupList | SiteDeviceList | SiteImsiDefinition | SiteIpDomainList | SiteMonitoring | SiteSimCardList | SiteSliceList | SiteSmallCellList | SiteUpfList | string | undefined;
}